backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.8.1
python-multipart==0.0.22
numpy==2.2.5
faiss-cpu==1.10.0
pytest==8.3.5
httpx==0.28.1
ruff==0.11.7
email-validator==2.3.0
